[Refactor][Elementwise] Read dtype at forward across the elementwise family (#1840)
Closes #1832.

24 elementwise constructors stop taking `dtype`; the op reads it from
the tensor at forward and caches one specialization per dtype.
`AlibiFwdOp` and `SinusoidalFwdOp` keep it — they generate a tensor from
their arguments, so no input supplies it.

### One way to reach a kernel

`_PerDtypeKernels._entry(dtype)` is the family's single lookup. A
subclass supplies `_build_entry` and gets a `KernelEntry` holding the
kernel, the dtype it was compiled for, and the resolved output dtype —
replacing a `_bool_storage` slot written as a side effect of building a
kernel, a constructor branch that skipped construction for integers, and
an `output_dtype_override` the manifest already answers.

### The op stops deciding for the backend

`Kernel.specialize(dtype) -> (impl, ctor_dtype)`, identity by default,
lets a backend say which implementation serves a semantic dtype and in
what storage. Every builder constructs what it returns. Gone from the op
layer: the `{op}_bool_storage` second slot on twelve ops, every
`torch.uint8`, the bool↔uint8 reinterpretation of operands and masks,
integer→float32 promotion for `reciprocal`, the 0-d scalar reshape in
`masked_fill`, and capability read from the family default instead of an
installed override.

A backend serving bool or integers natively declares nothing and is
constructed with the semantic dtype.
`test_builder_constructs_what_the_backend_specialized` injects a probe
backend across 22 builder shapes and requires that the probe is what got
built, with the dtype it asked for.

### Defects fixed

- `RoundFwdOp` read a stale `self.dtype`: an integer call followed by a
float call skipped rounding.
- `KernelEntry` was invisible to `Op.autotune`, so no built elementwise
kernel could be tuned.
- Compiling any unary op on a non-contiguous input failed a stride
assertion — the fake kept the input's layout while the real path returns
contiguous. Untested until now.
- `FusedGatedOp` committed new dimensions before validating the dtype.

### Behaviour change

Scalar parameters are baked into kernels and can only be checked against
an element type, so an out-of-range scalar raises at first use rather
than at construction.

### Deletions

`_FROZEN_UNARY_ACTIVATION_SIGNATURES`, a hand-copied snapshot of ten
constructor signatures that validator check C3 enforces from the
manifest. `compat_default` — a manifest field, validator branch, spec
paragraph and two tests existing so one op could take `num_experts=None`
and reject it later. 167 lines of `Args:` entries for parameters no
constructor accepts, with ruff's `DOC102` enabled to keep them out.

`pytest tests/ops/ -m smoke`: 2702 passed, 16 skipped.
`validate_manifest --strict`: clean.

+++ benchmarks/benchmark_base.py
@@ -520,7 +520,7 @@ class ManifestBenchmark(BenchmarkBase[Any]):
 
     Usage::
 
-        op = SumFwdOp(dtype=dtype, dim=0)
+        op = SumFwdOp(dim=0)
         bm = ManifestBenchmark("SumFwdOp", op, workload)
         result = bm.profile(op, *inputs)
     """
