+++ benchmarks/ops/bench_unary_elementwise.py
@@ -105,7 +105,7 @@ _EXP_OP = "ExpFwdOp"
 def test_exp_bench(shape: tuple, dtype: torch.dtype) -> None:
     inputs = draw_normal(shape, dtype)
     n_total = inputs[0].numel()
-    op = ExpFwdOp(N_total=n_total, dtype=dtype)
+    op = ExpFwdOp(N_total=n_total)
     bm = ManifestBenchmark(_EXP_OP, op, UnaryWorkload(shape, dtype))
     _profile_and_record(op, bm, inputs, torch.exp, {"shape": shape, "dtype": dtype, "n_total": n_total})
 
@@ -117,7 +117,7 @@ _LOG_OP = "LogFwdOp"
 def test_log_bench(shape: tuple, dtype: torch.dtype) -> None:
     inputs = draw_positive_away_from_zero(shape, dtype)
     n_total = inputs[0].numel()
-    op = LogFwdOp(N_total=n_total, dtype=dtype)
+    op = LogFwdOp(N_total=n_total)
     bm = ManifestBenchmark(_LOG_OP, op, UnaryWorkload(shape, dtype))
     _profile_and_record(op, bm, inputs, torch.log, {"shape": shape, "dtype": dtype, "n_total": n_total})
 
@@ -129,7 +129,7 @@ _SQRT_OP = "SqrtFwdOp"
 def test_sqrt_bench(shape: tuple, dtype: torch.dtype) -> None:
     inputs = draw_positive_away_from_zero(shape, dtype)
     n_total = inputs[0].numel()
-    op = SqrtFwdOp(N_total=n_total, dtype=dtype)
+    op = SqrtFwdOp(N_total=n_total)
     bm = ManifestBenchmark(_SQRT_OP, op, UnaryWorkload(shape, dtype))
     _profile_and_record(op, bm, inputs, torch.sqrt, {"shape": shape, "dtype": dtype, "n_total": n_total})
 
@@ -141,7 +141,7 @@ _RSQRT_OP = "RsqrtFwdOp"
 def test_rsqrt_bench(shape: tuple, dtype: torch.dtype) -> None:
     inputs = draw_positive_away_from_zero(shape, dtype)
     n_total = inputs[0].numel()
-    op = RsqrtFwdOp(N_total=n_total, dtype=dtype)
+    op = RsqrtFwdOp(N_total=n_total)
     bm = ManifestBenchmark(_RSQRT_OP, op, UnaryWorkload(shape, dtype))
     _profile_and_record(op, bm, inputs, torch.rsqrt, {"shape": shape, "dtype": dtype, "n_total": n_total})
 
@@ -153,7 +153,7 @@ _ABS_OP = "AbsFwdOp"
 def test_abs_bench(shape: tuple, dtype: torch.dtype) -> None:
     inputs = draw_normal(shape, dtype)
     n_total = inputs[0].numel()
-    op = AbsFwdOp(N_total=n_total, dtype=dtype)
+    op = AbsFwdOp(N_total=n_total)
     bm = ManifestBenchmark(_ABS_OP, op, UnaryWorkload(shape, dtype))
     _profile_and_record(op, bm, inputs, torch.abs, {"shape": shape, "dtype": dtype, "n_total": n_total})
 
@@ -165,7 +165,7 @@ _NEG_OP = "NegFwdOp"
 def test_neg_bench(shape: tuple, dtype: torch.dtype) -> None:
     inputs = draw_normal(shape, dtype)
     n_total = inputs[0].numel()
-    op = NegFwdOp(N_total=n_total, dtype=dtype)
+    op = NegFwdOp(N_total=n_total)
     bm = ManifestBenchmark(_NEG_OP, op, UnaryWorkload(shape, dtype))
     _profile_and_record(op, bm, inputs, torch.neg, {"shape": shape, "dtype": dtype, "n_total": n_total})
 
@@ -177,7 +177,7 @@ _RECIPROCAL_OP = "ReciprocalFwdOp"
 def test_reciprocal_bench(shape: tuple, dtype: torch.dtype) -> None:
     inputs = draw_positive_away_from_zero(shape, dtype)
     n_total = inputs[0].numel()
-    op = ReciprocalFwdOp(N_total=n_total, dtype=dtype)
+    op = ReciprocalFwdOp(N_total=n_total)
     bm = ManifestBenchmark(_RECIPROCAL_OP, op, UnaryWorkload(shape, dtype))
     _profile_and_record(op, bm, inputs, torch.reciprocal, {"shape": shape, "dtype": dtype, "n_total": n_total})
 
@@ -189,7 +189,7 @@ _SIGN_OP = "SignFwdOp"
 def test_sign_bench(shape: tuple, dtype: torch.dtype) -> None:
     inputs = draw_normal(shape, dtype)
     n_total = inputs[0].numel()
-    op = SignFwdOp(N_total=n_total, dtype=dtype)
+    op = SignFwdOp(N_total=n_total)
     bm = ManifestBenchmark(_SIGN_OP, op, UnaryWorkload(shape, dtype))
     _profile_and_record(op, bm, inputs, torch.sign, {"shape": shape, "dtype": dtype, "n_total": n_total})
 
@@ -201,7 +201,7 @@ _SIN_OP = "SinFwdOp"
 def test_sin_bench(shape: tuple, dtype: torch.dtype) -> None:
     inputs = draw_normal(shape, dtype)
     n_total = inputs[0].numel()
-    op = SinFwdOp(N_total=n_total, dtype=dtype)
+    op = SinFwdOp(N_total=n_total)
     bm = ManifestBenchmark(_SIN_OP, op, UnaryWorkload(shape, dtype))
     _profile_and_record(op, bm, inputs, torch.sin, {"shape": shape, "dtype": dtype, "n_total": n_total})
 
@@ -213,7 +213,7 @@ _COS_OP = "CosFwdOp"
 def test_cos_bench(shape: tuple, dtype: torch.dtype) -> None:
     inputs = draw_normal(shape, dtype)
     n_total = inputs[0].numel()
-    op = CosFwdOp(N_total=n_total, dtype=dtype)
+    op = CosFwdOp(N_total=n_total)
     bm = ManifestBenchmark(_COS_OP, op, UnaryWorkload(shape, dtype))
     _profile_and_record(op, bm, inputs, torch.cos, {"shape": shape, "dtype": dtype, "n_total": n_total})
 
@@ -225,7 +225,7 @@ _FLOOR_OP = "FloorFwdOp"
 def test_floor_bench(shape: tuple, dtype: torch.dtype) -> None:
     inputs = draw_normal(shape, dtype)
     n_total = inputs[0].numel()
-    op = FloorFwdOp(N_total=n_total, dtype=dtype)
+    op = FloorFwdOp(N_total=n_total)
     bm = ManifestBenchmark(_FLOOR_OP, op, UnaryWorkload(shape, dtype))
     _profile_and_record(op, bm, inputs, torch.floor, {"shape": shape, "dtype": dtype, "n_total": n_total})
 
@@ -237,7 +237,7 @@ _CEIL_OP = "CeilFwdOp"
 def test_ceil_bench(shape: tuple, dtype: torch.dtype) -> None:
     inputs = draw_normal(shape, dtype)
     n_total = inputs[0].numel()
-    op = CeilFwdOp(N_total=n_total, dtype=dtype)
+    op = CeilFwdOp(N_total=n_total)
     bm = ManifestBenchmark(_CEIL_OP, op, UnaryWorkload(shape, dtype))
     _profile_and_record(op, bm, inputs, torch.ceil, {"shape": shape, "dtype": dtype, "n_total": n_total})
 
@@ -249,7 +249,7 @@ _ROUND_OP = "RoundFwdOp"
 def test_round_bench(shape: tuple, dtype: torch.dtype) -> None:
     inputs = draw_normal(shape, dtype)
     n_total = inputs[0].numel()
-    op = RoundFwdOp(N_total=n_total, dtype=dtype)
+    op = RoundFwdOp(N_total=n_total)
     bm = ManifestBenchmark(_ROUND_OP, op, UnaryWorkload(shape, dtype))
     _profile_and_record(op, bm, inputs, torch.round, {"shape": shape, "dtype": dtype, "n_total": n_total})
 
@@ -261,7 +261,7 @@ _TRUNC_OP = "TruncFwdOp"
 def test_trunc_bench(shape: tuple, dtype: torch.dtype) -> None:
     inputs = draw_normal(shape, dtype)
     n_total = inputs[0].numel()
-    op = TruncFwdOp(N_total=n_total, dtype=dtype)
+    op = TruncFwdOp(N_total=n_total)
     bm = ManifestBenchmark(_TRUNC_OP, op, UnaryWorkload(shape, dtype))
     _profile_and_record(op, bm, inputs, torch.trunc, {"shape": shape, "dtype": dtype, "n_total": n_total})
 
@@ -273,7 +273,7 @@ _ERF_OP = "ErfFwdOp"
 def test_erf_bench(shape: tuple, dtype: torch.dtype) -> None:
     inputs = draw_normal(shape, dtype)
     n_total = inputs[0].numel()
-    op = ErfFwdOp(N_total=n_total, dtype=dtype)
+    op = ErfFwdOp(N_total=n_total)
     bm = ManifestBenchmark(_ERF_OP, op, UnaryWorkload(shape, dtype))
     _profile_and_record(op, bm, inputs, torch.erf, {"shape": shape, "dtype": dtype, "n_total": n_total})
 
@@ -285,7 +285,7 @@ _LOG1P_OP = "Log1pFwdOp"
 def test_log1p_bench(shape: tuple, dtype: torch.dtype) -> None:
     inputs = draw_positive_away_from_zero(shape, dtype)
     n_total = inputs[0].numel()
-    op = Log1pFwdOp(N_total=n_total, dtype=dtype)
+    op = Log1pFwdOp(N_total=n_total)
     bm = ManifestBenchmark(_LOG1P_OP, op, UnaryWorkload(shape, dtype))
     _profile_and_record(op, bm, inputs, torch.log1p, {"shape": shape, "dtype": dtype, "n_total": n_total})
 
@@ -297,7 +297,7 @@ _EXPM1_OP = "Expm1FwdOp"
 def test_expm1_bench(shape: tuple, dtype: torch.dtype) -> None:
     inputs = draw_normal(shape, dtype)
     n_total = inputs[0].numel()
-    op = Expm1FwdOp(N_total=n_total, dtype=dtype)
+    op = Expm1FwdOp(N_total=n_total)
     bm = ManifestBenchmark(_EXPM1_OP, op, UnaryWorkload(shape, dtype))
     _profile_and_record(op, bm, inputs, torch.expm1, {"shape": shape, "dtype": dtype, "n_total": n_total})
 
@@ -313,7 +313,7 @@ _LOGICAL_NOT_OP = "LogicalNotFwdOp"
 def test_logical_not_bench(shape: tuple, dtype: torch.dtype) -> None:
     inputs = draw_bool(shape, dtype)
     n_total = inputs[0].numel()
-    op = LogicalNotFwdOp(N_total=n_total, dtype=dtype)
+    op = LogicalNotFwdOp(N_total=n_total)
     bm = ManifestBenchmark(_LOGICAL_NOT_OP, op, UnaryWorkload(shape, dtype))
     _profile_and_record(op, bm, inputs, torch.logical_not, {"shape": shape, "dtype": dtype, "n_total": n_total})
 
@@ -325,7 +325,7 @@ _BITWISE_NOT_OP = "BitwiseNotFwdOp"
 def test_bitwise_not_bench(shape: tuple, dtype: torch.dtype) -> None:
     inputs = draw_int(shape, dtype)
     n_total = inputs[0].numel()
-    op = BitwiseNotFwdOp(N_total=n_total, dtype=dtype)
+    op = BitwiseNotFwdOp(N_total=n_total)
     bm = ManifestBenchmark(_BITWISE_NOT_OP, op, UnaryWorkload(shape, dtype))
     _profile_and_record(op, bm, inputs, torch.bitwise_not, {"shape": shape, "dtype": dtype, "n_total": n_total})
 
@@ -337,7 +337,7 @@ _ISNAN_OP = "IsnanFwdOp"
 def test_isnan_bench(shape: tuple, dtype: torch.dtype) -> None:
     inputs = draw_special_floats(shape, dtype)
     n_total = inputs[0].numel()
-    op = IsnanFwdOp(N_total=n_total, dtype=dtype)
+    op = IsnanFwdOp(N_total=n_total)
     bm = ManifestBenchmark(_ISNAN_OP, op, UnaryWorkload(shape, dtype))
     _profile_and_record(op, bm, inputs, torch.isnan, {"shape": shape, "dtype": dtype, "n_total": n_total})
 
@@ -349,7 +349,7 @@ _ISINF_OP = "IsinfFwdOp"
 def test_isinf_bench(shape: tuple, dtype: torch.dtype) -> None:
     inputs = draw_special_floats(shape, dtype)
     n_total = inputs[0].numel()
-    op = IsinfFwdOp(N_total=n_total, dtype=dtype)
+    op = IsinfFwdOp(N_total=n_total)
     bm = ManifestBenchmark(_ISINF_OP, op, UnaryWorkload(shape, dtype))
     _profile_and_record(op, bm, inputs, torch.isinf, {"shape": shape, "dtype": dtype, "n_total": n_total})
 
@@ -361,7 +361,7 @@ _ISFINITE_OP = "IsfiniteFwdOp"
 def test_isfinite_bench(shape: tuple, dtype: torch.dtype) -> None:
     inputs = draw_special_floats(shape, dtype)
     n_total = inputs[0].numel()
-    op = IsfiniteFwdOp(N_total=n_total, dtype=dtype)
+    op = IsfiniteFwdOp(N_total=n_total)
     bm = ManifestBenchmark(_ISFINITE_OP, op, UnaryWorkload(shape, dtype))
     _profile_and_record(op, bm, inputs, torch.isfinite, {"shape": shape, "dtype": dtype, "n_total": n_total})
